// packages/protocol/typescript/generated/yiru/runtime/v1/terminal_pb.d.ts
export declare type TerminalServiceOpenMultiplexRequest = Message<"yiru.runtime.v1.TerminalServiceOpenMultiplexRequest"> & {
  /**
   * @generated from field: string client_instance_id = 1;
   */
  clientInstanceId: string;

  /**
   * @generated from field: string environment_id = 2;
   */
  environmentId: string;
};

/**
 * Describes the message yiru.runtime.v1.TerminalServiceOpenMultiplexRequest.
 * Use `create(TerminalServiceOpenMultiplexRequestSchema)` to create a new message.
 */
export declare const TerminalServiceOpenMultiplexRequestSchema: GenMessage<TerminalServiceOpenMultiplexRequest>;

/**
 * @generated from message yiru.runtime.v1.TerminalServiceOpenMultiplexResponse
 */
export declare type TerminalServiceOpenMultiplexResponse = Message<"yiru.runtime.v1.TerminalServiceOpenMultiplexResponse"> & {
  /**
   * @generated from field: string bulk_ticket = 2;
   */
  bulkTicket: string;

  /**
   * @generated from field: uint64 expires_at = 3;
   */
  expiresAt: bigint;

  /**
   * @generated from field: uint32 max_frame_bytes = 4;
   */
  maxFrameBytes: number;
};

/**
 * Describes the message yiru.runtime.v1.TerminalServiceOpenMultiplexResponse.
 * Use `create(TerminalServiceOpenMultiplexResponseSchema)` to create a new message.
 */
export declare const TerminalServiceOpenMultiplexResponseSchema: GenMessage<TerminalServiceOpenMultiplexResponse>;